Multi-agent, LangChain
LLMs like Open AI or LLama2 gives us text based outputs and instructions, But they cannot perform any actions. Whereas agents perform any task based on the instructions using ReAct, ChainOfThought, ChainOfProgram. A agent breaks down the task and understands the task and then tries to execute it. Some of the tasks that can be performed are searching through various search engines for additional information, accessing power shell to execute commands, reading csv files and analysing and generating charts and reports. With tools like Connary, we can mail the generated reports and insights to Email /save in a file in local system.

This project presents various tools that can be used from lang chain. it also shows the advantages of agents, multi-agent architecture and ways to optimise and reduce number of calls to LLM models.

The main objective of this project is to perform any task with just a simple prompt. For instance, A simple prompt like "read a file/perform data analysis on a file_name.txt/csv and present the results in a chart". This can be executed using combinations of LLMs and Multi agent architecture.